Add files via upload
Mapping Rio's Favelas by Emily Matchar discusses mapping the existing favelas that were not captured on the traditional satellite systems such as Google and other online platforms. Accurately capturing the community's presence on these online maps because it provides data to examine the social and environmental issues of the geographic region. Without any map presence, the community can be easily overlooked. As part of the solution during the last Olympics, Google partnered with a Brazilian nonprofit, AfroReggae to properly map the favelas with local residents. The article did not talk about any employment or compensation details for the community Google worked with which links to the extractive nature of planning.

This article connected to me because I thought about the different ways mapping can significantly impact a community locally or globally. In the local geographic context, I wonder if there are any other American neighborhoods that face this same issue. This is important because I believe there is a strong need to represent data accurately because it can determine funding, future outlooks, etc. For example, Census data is significant in determining funding distributions so I further contemplate about overall connections between the gaps in mapping and data collection (Census).